backend/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.31,<0.32
httpx>=0.25,<0.27
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-jose[cryptography]>=3.3,<4
